requirements.txt
accelerate==1.3.0
africastalking==1.2.9
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
eralchemy==1.5.0
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Limiter==3.12
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
huggingface-hub==0.28.1
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
liboqs==0.9.1
limits==4.4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pqcryptography==0.0.3
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.22.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyoqs-sdk==1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.44.1
streamlit-autorefresh==1.0.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3